### 2019-02-27 d1f90be
moving figures to paper/ [skip ci]
