+++ paper.bib
@@ -1,175 +1,0 @@
-
-@article{blackwell1973,
-author = "Blackwell, David and MacQueen, James B.",
-doi = "10.1214/aos/1176342372",
-fjournal = "The Annals of Statistics",
-journal = "Ann. Statist.",
-month = "03",
-number = "2",
-pages = "353--355",
-publisher = "The Institute of Mathematical Statistics",
-title = "Ferguson Distributions Via Polya Urn Schemes",
-url = "https://doi.org/10.1214/aos/1176342372",
-volume = "1",
-year = "1973"
-}
-
-@article{Hastings1970,
- ISSN = {00063444},
- URL = {http://www.jstor.org/stable/2334940},
- author = {W. K. Hastings},
- journal = {Biometrika},
- number = {1},
- pages = {97--109},
- publisher = {[Oxford University Press, Biometrika Trust]},
- title = {Monte Carlo Sampling Methods Using Markov Chains and Their Applications},
- volume = {57},
- year = {1970}
-}
-
-@article{hannah2011,
-  title={Dirichlet process mixtures of generalized linear models},
-  author={Hannah, Lauren A and Blei, David M and Powell, Warren B},
-  journal={Journal of Machine Learning Research},
-  volume={12},
-  number={Jun},
-  pages={1923--1953},
-  year={2011}
-}
-
-@book{muller2015,
-  title={Bayesian Nonparametric Data Analysis},
-  author={M{\"u}ller, P. and Quintana, F.A. and Jara, A. and Hanson, T.},
-  isbn={9783319189673},
-  lccn={2015943065},
-  series={Springer Series in Statistics},
-  url={https://books.google.com/books?id=eH5ErgEACAAJ},
-  year={2015},
-  publisher={Springer International Publishing}
-}
-
-@article{Neal2000,
-author = { Radford M. Neal },
-title = {Markov Chain Sampling Methods for Dirichlet Process Mixture Models},
-journal = {Journal of Computational and Graphical Statistics},
-volume = {9},
-number = {2},
-pages = {249-265},
-year  = {2000},
-publisher = {Taylor & Francis},
-doi = {10.1080/10618600.2000.10474879},
-URL = {https://amstat.tandfonline.com/doi/abs/10.1080/10618600.2000.10474879},
-eprint = {https://amstat.tandfonline.com/doi/pdf/10.1080/10618600.2000.10474879}
-}
-
-@article{gershman2012,
-  added-at = {2015-05-22T11:42:38.000+0200},
-  author = {Gershman, Samuel J. and Blei, David M.},
-  biburl = {https://www.bibsonomy.org/bibtex/21dbad47ef24ea22daef95f1d79c062dc/jil},
-  doi = {http://dx.doi.org/10.1016/j.jmp.2011.08.004},
-  interhash = {60a992ffab2105467ae596dc57fb098f},
-  intrahash = {1dbad47ef24ea22daef95f1d79c062dc},
-  issn = {0022-2496},
-  journal = {Journal of Mathematical Psychology },
-  keywords = {bayesian blei models nonparametric nonparametrics tutorial},
-  number = 1,
-  pages = {1 - 12},
-  timestamp = {2015-05-22T11:42:38.000+0200},
-  title = {A tutorial on Bayesian nonparametric models },
-  url = {http://www.sciencedirect.com/science/article/pii/S002224961100071X},
-  volume = 56,
-  year = 2012
-}
-
-@article{ferguson1973,
-author = "Ferguson, Thomas S.",
-doi = "10.1214/aos/1176342360",
-fjournal = "The Annals of Statistics",
-journal = "Ann. Statist.",
-month = "03",
-number = "2",
-pages = "209--230",
[…]
