### 2019-02-27 #1283 @stablemarkets
Hi Daniel. I forgot to include a .png file that was called in paper.md. Maybe that was causing a compilation error? I fixed it now so that my paper.md, paper.bib, and the figure are all in the paper/ subdirectory of my GitHub repo. Maybe it will work now.
