### 2019-02-28 #1287 @whedon
[REVIEW]: ChiRP: Chinese Restaurant Process Mixtures for Regression and Clustering
**Submitting author:** @stablemarkets (<a href="http://orcid.org/0000-0002-0437-4611">Arman Oganisian</a>)
**Repository:** <a href="https://github.com/stablemarkets/ChiRP" target ="_blank">https://github.com/stablemarkets/ChiRP</a>
**Version:** 1.0.0
