### 2019-02-28 #1287 @stablemarkets
Hi @pjotrp. Thank you for working on this.

I've provided responses to several items in the check list below:

* [ ] **Repository:** Source code is located in the [R/](https://github.com/stablemarkets/ChiRP/tree/master/R) subdirectory.
* [ ]  **License:** MIT license is here: [LICENSE.md](https://github.com/stablemarkets/ChiRP/blob/master/LICENSE.md)
* [ ]  **Version:** See [DESCRIPTION](https://github.com/stablemarkets/ChiRP/blob/master/DESCRIPTION) file where we list "Version: 1.0.0".
* [ ]  **Authorship:** I am the sole author of the code base.

* [ ]  **Installation:** Installation is documented in the README.md file of the repo. Travis CI is used to confirm builds on Linux and OSX.
* [ ]  **Functionality:** This can be partially evaluated by replicating some of the examples in the package's [companion site](https://stablemarkets.github.io/ChiRPsite/examples.html).
* [ ]  **Performance:** I don't believe I've made any performance claims.

* [ ]  **A statement of need:** I feel this is addressed in the paper.md file but I will leave to reviewers/editors to judge.
* [ ]  **Installation instructions:** Dependency packages are listed in in (https://github.com/stablemarkets/ChiRP/blob/master/DESCRIPTION). Upon installing `ChiRP`, R automatically checks for these dependencies and installs if necessary
* [ ]  **Example usage:** Several usage examples are given on the package's [companion site](https://stablemarkets.github.io/ChiRPsite/examples.html)
* [ ]  **Functionality documentation:** See "Documentation and Examples" section of the repo README.md for instructions on how to access API help files within R.
* [ ]  **Automated tests:** The `testthat()`R package is used for automated testing. This is integrated with Travis CI so that tests are run upon every commit. Coveralls is used to track the coverage of these automated tests. Coverage is currently 92%.
* [ ]  **Community guidelines:** Issue reporting and contact info are given in the "Reporting Issues" and "Contact" sections of the repo README.md
