### 2019-02-28 277d916
adding JOSS badge to README [skip ci]
