### 2019-02-28 7f22d23
adding JOSS badge to README [skip ci]
