### 2019-03-07 #1287 @agisga
Hi!

I have reviewed this submission. Generally it is pretty good. My review is given below.

**Summary:**

This submission is an R package that implements several Bayesian models from the family of Chinese Restaurant Process (CRP) mixtures. The methods implemented in ChiRP can be used for regression, binary classification, clustering, and related inference tasks. These are very common data analysis tasks, which one commonly encounters in many scientific disciplines. However, the author seems to be motivated specifically by biomedical applications. The advantages of the ChiRP models against alternative methods, include their nonparametric nature, the ability to return interval estimates, and to obtain posterior distributions for predictions as well as for cluster assignments on training and test data. In particular, unlike many other clustering algorithms, here the number of clusters is not specified a priori, and is determined automatically from the data (although an initial number of clusters needs to be given to kick off the MCMC sampler). The package website including model description and examples is really informative and great (in addition, the R Shiny tutorial based on Dirichlet Processes is just awesome).
The R code is well-documented, and includes automated tests too. Code quality seems to be good (based on a very quick look through each of the source code files). I'm not an expert on the methodology but, running and tweaking the provided examples, the package code seems to be doing what it's supposed to do.
Some minor issues are listed below.

**Issues:**

* *One unchecked check box:* "Do all archival references that should have a DOI list one (e.g., papers, datasets, software)" ~~> No, the majority of cited papers don't have DOI specified.
* There are Acknowledgements, mentioning several individuals, at <https://stablemarkets.github.io/ChiRPsite/index.html> but not in the paper.
* I think that many people who have never heard of the "Chinese Restaurant Process (CRP)" may find this R package very useful, especially for clustering. Therefore, it would be good to cite a canonical reference for Chinese Restaurant Process (CRP) mixtures in the very beginning of the paper for those readers unfamiliar with this methodology.
* There are community guidelines for seeking support and reporting issues, but there is no "contribution" information, which is point 1) of the "Community Guidelines" check box.
* I get three warnings from `testthat` when running the automated tests. Not really an issue, but it probably would be better to not have any warning. (`test_ndp.r:23: warning: Test Init K Errors`, `test_pdp.r:25: warning: Test Init K Errors`, and `test_zdp.r:28: warning: Test Init K Errors`).
* Paper: "ARCHIVE" link broken in the PDF (not sure if that's an issue).
* Typos:
    - File name `./R/helper_funtions_pdp.R` ~~> might want to change to `./R/helper_functions_pdp.R` ("functions" spelled with a "c")
    - There are a couple of minor spelling mistakes on the project webpage, which can be found with a simple spell check.
    - Some of the math notation is not displayed correctly in the bullet point list following "Package implementation details:" at https://stablemarkets.github.io/ChiRPsite/modeldesc.html

None of these are serious issues.
