### 2019-03-07 #1287 @pjotrp
Thank you @agisga! @stablemarkets if you can address these points quickly we can publish!
