### 2019-03-07 e225f7d
implementing JOSS reviewer comments
