+++ paper/paper.bib
@@ -1,32 +1,3 @@
-
-@article{blackwell1973,
-author = "Blackwell, David and MacQueen, James B.",
-doi = "10.1214/aos/1176342372",
-fjournal = "The Annals of Statistics",
-journal = "Ann. Statist.",
-month = "03",
-number = "2",
-pages = "353--355",
-publisher = "The Institute of Mathematical Statistics",
-title = "Ferguson Distributions Via Polya Urn Schemes",
-url = "https://doi.org/10.1214/aos/1176342372",
-volume = "1",
-year = "1973"
-}
-
-@article{Hastings1970,
- ISSN = {00063444},
- URL = {http://www.jstor.org/stable/2334940},
- author = {W. K. Hastings},
- journal = {Biometrika},
- number = {1},
- pages = {97--109},
- publisher = {[Oxford University Press, Biometrika Trust]},
- title = {Monte Carlo Sampling Methods Using Markov Chains and Their Applications},
- volume = {57},
- year = {1970}
-}
-
 @article{hannah2011,
   title={Dirichlet process mixtures of generalized linear models},
   author={Hannah, Lauren A and Blei, David M and Powell, Warren B},
@@ -34,7 +5,8 @@ year = "1973"
   volume={12},
   number={Jun},
   pages={1923--1953},
-  year={2011}
+  year={2011},
+  url ={http://www.jmlr.org/papers/volume12/hannah11a/hannah11a.pdf}
 }
 
 @book{muller2015,
@@ -45,7 +17,8 @@ year = "1973"
   series={Springer Series in Statistics},
   url={https://books.google.com/books?id=eH5ErgEACAAJ},
   year={2015},
-  publisher={Springer International Publishing}
+  publisher={Springer International Publishing},
+  doi={10.1007/978-3-319-18968-0}
 }
 
 @article{Neal2000,
@@ -78,22 +51,8 @@ eprint = {https://amstat.tandfonline.com/doi/pdf/10.1080/10618600.2000.10474879}
   title = {A tutorial on Bayesian nonparametric models },
   url = {http://www.sciencedirect.com/science/article/pii/S002224961100071X},
   volume = 56,
-  year = 2012
-}
-
-@article{ferguson1973,
-author = "Ferguson, Thomas S.",
-doi = "10.1214/aos/1176342360",
-fjournal = "The Annals of Statistics",
-journal = "Ann. Statist.",
-month = "03",
-number = "2",
-pages = "209--230",
-publisher = "The Institute of Mathematical Statistics",
-title = "A Bayesian Analysis of Some Nonparametric Problems",
-url = "https://doi.org/10.1214/aos/1176342360",
-volume = "1",
-year = "1973"
+  year = 2012,
+  doi={10.1016/j.jmp.2011.08.004}
 }
 
 @article{Hastings1970,
@@ -107,7 +66,8 @@ year = "1973"
  publisher = {[Oxford University Press, Biometrika Trust]},
  title = {Monte Carlo Sampling Methods Using Markov Chains and Their Applications},
  volume = {57},
- year = {1970}
+ year = {1970},
+ doi = {10.2307/2334940}
 }
 
 @ARTICLE{Oganisian2018,
@@ -123,23 +83,6 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@article{Metropolis1953,
-author = {Metropolis,Nicholas  and Rosenbluth,Arianna W.  and Rosenbluth,Marshall N.  and Teller,Augusta H.  and Teller,Edward },
-title = {Equation of State Calculations by Fast Computing Machines},
[…]
