### 2019-03-07 b8f6c30
edits to README per JOSS review comments
