### 2019-03-07 2c96312
Edits to error_check.r per JOSS reviewer comments
