### 2019-03-07 c719843
minor changes to paper.md [skip ci]
