### 2019-03-07 #1287 @stablemarkets
Hi @pjotrp and @agisga. Thank you for your thorough review! Thanks especially for reviewing the companion web site - extremely helpful.

The package repository has been updated with edits that address your feedback (see below for details). I think the paper.md file will need to be recompiled for the edits to be visible?

1. I added DOIs to all but two references in paper.bib. I could not find DOIs for two of the papers. One is an ArXiv paper (currently unpublished) so I think it has no DOI yet. For both papers, URLs are provided. Hope this is okay.
2. Thanks! I've added an acknowledgements sections to the paper mentioning these individuals.
3. I've added two citations to seminal CRP papers - Ferguson 1973, Blackwell 1973 - to the first paragraph when introducing CRPs.
4. Contribution instructions have been added to both the ChiRP [repository](https://stablemarkets.github.io/ChiRPsite/index.html) and the [companion site](https://github.com/stablemarkets/ChiRP). Users can contribute by proposing modifications to the base code or by adding usage examples to the companion site.
5. testthat warnings: Thanks for catching this! I looked into it. The warning is triggered when someone specifies and `init_k` with a length greater than 1. In `error_checks()`, I run a single check that returns an error if `init_k` is not numeric, or if `init_k` is not a scalar (e.g. length 2), or if `init_k<0`. If `init_k` is of length 2, an error is indeed returned so the test is passed. However, evaluating `init_k<0`, yields a length 2 logical vector while the other two conditions in the check yield length 1 logicals. So it shoots a warning saying it will only use the first element of `init_k<0` to evaluate the entire check. I fixed this by doing a nested check: first making sure `init_k` is numeric and length 1. Then, if that is passed, check to make sure if `init_k>0`. If it's not, throw an error message at the user. This passes the automated test without shooting an additional warning since `init_k<0`is evaluated only when we are sure `init_k` is a scalar.

6. I'm not sure but the ARCHIVE link is probably not working since I haven't archived the paper yet. According to [review process](https://joss.readthedocs.io/en/latest/submitting.html#the-review-process) it seems like we only archive after the review process is completed?
7. Thanks for catching these typos: I've renamed helper/functions_pdp.r, fixed the notation rendering on the site, as well as spell-checked each of the three pages on the companion site.
