### 2019-03-08 #1287 @pjotrp
Excellent. @agisga can you run through those points and confirm you are happy?
