### 2019-03-10 #1287 @agisga
@pjotrp

Everything looks good to me!

I don't think there is a new article proof. So, can't check that. But I have looked at all new commits in the ChiRP repo, and also confirmed that the automated tests now run without warnings.

I don't know how strict JOSS is about having DOIs for every article reference, but in my opinion it's fine to have those two DOIs missing, especially because URLs are provided. Thus, I have checked the last checkbox. I leave the final decision up to you.
