### 2019-03-12 #1287 @stablemarkets
Hi @pjotrp,

I confirm all references in my bibliography have a DOI where available. I've deposited my repository (including revisions discussed here) to Zenodo. Here is the DOI badge:

[![DOI](https://zenodo.org/badge/168850653.svg)](https://zenodo.org/badge/latestdoi/168850653)
